Add RowJsonSerializer and tests
